Aggregate climate data for different frequencies

Aggregate time series downloaded climate data to day, month or year. Only observations under the tags TSSM_CON, TMN_CON, TMX_CON, PTPM_CON, and BSHG_CON can be aggregated, since are the ones where methodology for aggregation is explicitly provided from the source

aggregate_climate(climate_data, frequency)

Arguments

climate_data

data.frame obtained from download functions. Only observations under the same tag can be aggregated

frequency

character with the aggregation frequency. ("day", "month" or "year")

Value

data.frame object with the aggregated data

Examples

lat <- c(5.166278, 5.166278, 4.982247, 4.982247, 5.166278)
lon <- c(-75.678072, -75.327859, -75.327859, -75.678072, -75.678072)
polygon <- sf::st_polygon(x = list(cbind(lon, lat)))
geometry <- sf::st_sfc(polygon)
roi <- sf::st_as_sf(geometry)
tssm <- download_climate_geom(roi, "2021-11-14", "2021-11-20", "TSSM_CON")
#> Original data is provided by the Institute of Hydrology, Meteorology
#> and Environmental Studies(IDEAM).
#> Reformatted by the package authors.
#> Stored and redistributed by Universidad de Los Andes under the Epiverse
#> TRACE iniative.
daily_tssm <- aggregate_climate(tssm, "day")