Download data dictionary for geospatial data

Download data dictionary for geospatial data

Usage

dictionary(dataset)

Arguments

dataset

character with the dataset name

Value

tibble with data dictionary

Examples

dictionary("DANE_MGNCNPV_2018_SETU")
#>      variable         tipo longitud
#> 1  DPTO_CCDGO         Text        2
#> 2  MPIO_CCDGO         Text        3
#> 3  MPIO_CDPMP         Text        5
#> 4  CLAS_CCDGO         Text        1
#> 5  SETR_CCDGO         Text        3
#> 6  SETR_CCNCT         Text        9
#> 7  SECR_CCDGO         Text        2
#> 8  SECR_CCNCT         Text       11
#> 9    ZU_CCDGO         Text        3
#> 10   ZU_CDIVI         Text        8
#> 11 SETU_CCDGO         Text        4
#> 12 SETU_CCNCT         Text       18
#> 13  DATOS_ANM         Text       50
#> 14    VERSION Long Integer       NA
#> 15       AREA       Double       NA
#> 16    LATITUD       Double       NA
#> 17   LONGITUD       Double       NA
#> 18 STCTNENCUE       Double       NA
#> 19  STP3_1_SI       Double       NA
#> 20  STP3_2_NO       Double       NA
#> 21   STP3A_RI       Double       NA
#> 22  STP3B_TCN       Double       NA
#> 23  STP4_1_SI       Double       NA
#> 24  STP4_2_NO       Double       NA
#> 25 STP9_1_USO       Double       NA
#> 26 STP9_2_USO       Double       NA
#> 27 STP9_3_USO       Double       NA
#> 28 STP9_4_USO       Double       NA
#> 29 STP9_2_1_M       Double       NA
#> 30 STP9_2_2_M       Double       NA
#> 31 STP9_2_3_M       Double       NA
#> 32 STP9_2_4_M       Double       NA
#> 33 STP9_2_9_M       Double       NA
#> 34 STP9_3_1_N       Double       NA
#> 35 STP9_3_2_N       Double       NA
#> 36 STP9_3_3_N       Double       NA
#> 37 STP9_3_4_N       Double       NA
#> 38 STP9_3_5_N       Double       NA
#> 39 STP9_3_6_N       Double       NA
#> 40 STP9_3_7_N       Double       NA
#> 41 STP9_3_8_N       Double       NA
#> 42 STP9_3_9_N       Double       NA
#> 43  STP9_3_10       Double       NA
#> 44  STP9_3_99       Double       NA
#> 45 STVIVIENDA       Double       NA
#> 46 STP14_1_TI       Double       NA
#> 47 STP14_2_TI       Double       NA
#> 48 STP14_3_TI       Double       NA
#> 49 STP14_4_TI       Double       NA
#> 50 STP14_5_TI       Double       NA
#> 51 STP14_6_TI       Double       NA
#> 52 STP15_1_OC       Double       NA
#> 53 STP15_2_OC       Double       NA
#> 54 STP15_3_OC       Double       NA
#> 55 STP15_4_OC       Double       NA
#> 56  TSP16_HOG       Double       NA
#> 57 STP19_EC_1       Double       NA
#> 58 STP19_ES_2       Double       NA
#> 59 STP19_EE_1       Double       NA
#> 60 STP19_EE_2       Double       NA
#> 61 STP19_EE_3       Double       NA
#> 62 STP19_EE_4       Double       NA
#> 63 STP19_EE_5       Double       NA
#> 64 STP19_EE_6       Double       NA
#> 65 STP19_EE_9       Double       NA
#> 66 STP19_ACU1       Double       NA
#> 67 STP19_ACU2       Double       NA
#> 68 STP19_ALC1       Double       NA
#> 69 STP19_ALC2       Double       NA
#> 70 STP19_GAS1       Double       NA
#> 71 STP19_GAS2       Double       NA
#> 72 STP19_GAS9       Double       NA
#> 73 STP19_REC1       Double       NA
#> 74 STP19_REC2       Double       NA
#> 75 STP19_INT1       Double       NA
#> 76 STP19_INT2       Double       NA
#> 77 STP19_INT9       Double       NA
#> 78 STP27_PERS       Double       NA
#> 79 STPERSON_L       Double       NA
#> 80 STPERSON_S       Double       NA
#> 81 STP32_1_SE       Double       NA
#> 82 STP32_2_SE       Double       NA
#> 83 STP34_1_ED       Double       NA
#> 84 STP34_2_ED       Double       NA
#> 85 STP34_3_ED       Double       NA
#> 86 STP34_4_ED       Double       NA
#> 87 STP34_5_ED       Double       NA
#> 88 STP34_6_ED       Double       NA
#> 89 STP34_7_ED       Double       NA
#> 90 STP34_8_ED       Double       NA
#> 91 STP34_9_ED       Double       NA
#> 92 STP51_PRIM       Double       NA
#> 93 STP51_SECU       Double       NA
#> 94 STP51_SUPE       Double       NA
#> 95 STP51_POST       Double       NA
#> 96 STP51_13_E       Double       NA
#> 97 STP51_99_E       Double       NA
#> 98       <NA>       Double       NA
#>                                                                                                            descripcion
#> 1                                                                                              Código del departamento
#> 2                                                                                                 Código del municipio
#> 3                                                                       Código concatenado que identifica al municipio
#> 4                                   Código de la clase 1 cabecera municipal, 2 centro poblado y 3 área resto municipal
#> 5                                                                                               Código de sector rural
#> 6                                   Código de sector rural concatenado (departamento, municipio, clase y sector rural)
#> 7                                                                                              Código de sección rural
#> 8                 Código de sección rural  concatenado (departamento, municipio, clase, sector rural y  sección rural)
#> 9                                                             Código zona urbana (cabecera municipal y centro poblado)
#> 10                                             Código zona urbana  concatenado (departamento, municipio y zona urbana)
#> 11                                                                                                Código sector urbano
#> 12 Código sector concatenado (departamento, municipio, clase, sector rural, sección rural, zona urbana y sector rural)
#> 13                                                                                             Nombre capa anonimizada
#> 14                                                                                    Año de la información geográfica
#> 15                   Área del sector urbano en metros cuadrados  (Sistema de coordenadas planas MAGNA_Colombia_Bogota)
#> 16                                                                   Coordenada de latitud centroide del sector urbano
#> 17                                                                  Coordenada de longitud centroide del sector urbano
#> 18                                                                                     Cantidad de Encuestas CNPV 2018
#> 19                                                     Cantidad de encuestas que reportaron estar en territorio étnico
#> 20                                                  Cantidad de encuestas que reportaron no estar en territorio étnico
#> 21                                  Cantidad de encuestas que reportaron estar en territorio étnico Resguardo indígena
#> 22          Cantidad de encuestas que reportaron estar en territorio étnico Territorio colectivo de comunidades negras
#> 23                                                      Cantidad de encuestas que reportaron estar en áreas protegidas
#> 24                                                   Cantidad de encuestas que reportaron no estar en áreas protegidas
#> 25                                                                                 Conteo de unidades con uso vivienda
#> 26                                                                                    Conteo de unidades con uso mixto
#> 27                                                                           Conteo de unidades con uso no residencial
#> 28                                                                                      Conteo de unidades con uso LEA
#> 29                                                          Conteo de unidades mixtas con uso no residencial industria
#> 30                                                           Conteo de unidades mixtas con uso no residencial comercio
#> 31                                                          Conteo de unidades mixtas con uso no residencial servicios
#> 32                             Conteo de unidades mixtas con uso no residencial agropecuario, agroindustrial, forestal
#> 33                                                    Conteo de unidades mixtas con uso no residencial sin información
#> 34                                                               Conteo de unidades no residénciales con uso Industria
#> 35                                                                Conteo de unidades no residénciales con uso Comercio
#> 36                                                               Conteo de unidades no residénciales con uso Servicios
#> 37                                  Conteo de unidades no residénciales con uso Agropecuario, Agroindustrial, Forestal
#> 38                                                           Conteo de unidades no residénciales con uso Institucional
#> 39                                          Conteo de unidades no residénciales con uso Lote (Unidad sin construcción)
#> 40                                                       Conteo de unidades no residénciales con uso Parque/Zona Verde
#> 41                                                       Conteo de unidades no residénciales con uso Minero-Energético
#> 42                                       Conteo de unidades no residénciales con uso Protección/Conservación ambiental
#> 43                                                         Conteo de unidades no residénciales con uso En Construcción
#> 44                                                         Conteo de unidades no residénciales con uso Sin información
#> 45                                                                                                 Conteo de viviendas
#> 46                                                                                       Conteo de viviendas tipo Casa
#> 47                                                                                Conteo de viviendas tipo Apartamento
#> 48                                                                                     Conteo de viviendas tipo cuarto
#> 49                                                              Conteo de viviendas tipo Vivienda tradicional Indígena
#> 50                                 Conteo de viviendas tipo Vivienda tradicional étnica (Afrocolombiana, Isleña, Rrom)
#> 51                 Conteo de viviendas tipo Otro (contenedor, carpa, embarcación, vagón, cueva, refugio natural, etc.)
#> 52                                                                  Conteo de viviendas Ocupada con personas presentes
#> 53                                                         Conteo de viviendas Ocupada con todas las personas ausentes
#> 54                                              Conteo de viviendas Vivienda  temporal (para vacaciones, trabajo etc.)
#> 55                                                                                      Conteo de viviendas Desocupada
#> 56                                                                                                   Conteo de hogares
#> 57                                                                           Conteo de viviendas con energía eléctrica
#> 58                                                                           Conteo de viviendas sin energía eléctrica
#> 59                              Conteo de viviendas que reportan recibir facturación de energía eléctrica en Estrato 1
#> 60                              Conteo de viviendas que reportan recibir facturación de energía eléctrica en Estrato 2
#> 61                              Conteo de viviendas que reportan recibir facturación de energía eléctrica en Estrato 3
#> 62                              Conteo de viviendas que reportan recibir facturación de energía eléctrica en Estrato 4
#> 63                              Conteo de viviendas que reportan recibir facturación de energía eléctrica en Estrato 5
#> 64                              Conteo de viviendas que reportan recibir facturación de energía eléctrica en Estrato 6
#> 65             Conteo de viviendas que reportan recibir facturación de energía eléctrica en No sabe o no tiene estrato
#> 66                                                                       Conteo de viviendas con servicio de acueducto
#> 67                                                                       Conteo de viviendas sin servicio de acueducto
#> 68                                                                  Conteo de viviendas con servicio de alcantarillado
#> 69                                                                  Conteo de viviendas sin servicio de alcantarillado
#> 70                                             Conteo de viviendas con servicio de gas natural conectado a red pública
#> 71                                             Conteo de viviendas sin servicio de gas natural conectado a red pública
#> 72                              Conteo de viviendas sin información de servicio de gas natural conectado a red pública
#> 73                                                          Conteo de viviendas con servicio de recolección de basuras
#> 74                                                          Conteo de viviendas sin servicio de recolección de basuras
#> 75                                                                        Conteo de viviendas con servicio de internet
#> 76                                                                        Conteo de viviendas sin servicio de internet
#> 77                                                         Conteo de viviendas sin información de servicio de internet
#> 78                                                                                                  Número de personas
#> 79                                                                                          Conteo de personas en LEAS
#> 80                                                                          Conteo de personas en hogares particulares
#> 81                                                                                                   Conteo de hombres
#> 82                                                                                                   Conteo de mujeres
#> 83                                                                                 Conteo de personas entre 0 - 9 años
#> 84                                                                               Conteo de personas entre 10 - 19 años
#> 85                                                                               Conteo de personas entre 20 - 29 años
#> 86                                                                               Conteo de personas entre 30 - 39 años
#> 87                                                                               Conteo de personas entre 40 - 49 años
#> 88                                                                               Conteo de personas entre 50 - 59 años
#> 89                                                                               Conteo de personas entre 60 - 69 años
#> 90                                                                               Conteo de personas entre 70 - 79 años
#> 91                                                                             Conteo de personas entre 80  y más años
#> 92   Conteo de personas donde el nivel educativo del último año alcanzado es Preescolar - Pre jardín Y Básica primaria
#> 93                     Conteo de personas donde el nivel educativo del último año alcanzado es Básica secundaria 3...6
#> 94            Conteo de personas donde el nivel educativo del último año alcanzado es Técnica profesional 1 año  7...9
#> 95    Conteo de personas donde el nivel educativo del último año alcanzado es Especialización 1 año…maestría doctorado
#> 96                                     Conteo de personas donde el nivel educativo del último año alcanzado es Ninguno
#> 97                             Conteo de personas donde el nivel educativo del último año alcanzado es Sin información
#> 98                                                                                                                <NA>
#>                                                                           categoria_original
#> 1                                                                                       <NA>
#> 2                                                                                       <NA>
#> 3                                                                                       <NA>
#> 4                                                                                       <NA>
#> 5                                                                                       <NA>
#> 6                                                                                       <NA>
#> 7                                                                                       <NA>
#> 8                                                                                       <NA>
#> 9                                                                                       <NA>
#> 10                                                                                      <NA>
#> 11                                                                                      <NA>
#> 12                                                                                      <NA>
#> 13                                                                                      <NA>
#> 14                                                                                      <NA>
#> 15                                                                                      <NA>
#> 16                                                                                      <NA>
#> 17                                                                                      <NA>
#> 18                                                                                      <NA>
#> 19                                                                                      <NA>
#> 20                                                                                      <NA>
#> 21                                                                        Resguardo Indígena
#> 22                                                                                      TCCN
#> 23                                                                            Área protegida
#> 24                                                                                      <NA>
#> 25                                                                                  Vivienda
#> 26 Mixto (Espacio independiente y separado que combina vivienda con otro uso no residencial)
#> 27     Unidad NO Residencial (Espacio independiente y separado con uso diferente a vivienda)
#> 28                                                       Lugar especial de alojamiento (LEA)
#> 29                                                                                 Industria
#> 30                                                                                  Comercio
#> 31                                                                                 Servicios
#> 32                                                    Agropecuario, Agroindustrial, Forestal
#> 33                                                                           Sin información
#> 34                                                                                 Industria
#> 35                                                                                  Comercio
#> 36                                                                                 Servicios
#> 37                                                    Agropecuario, Agroindustrial, Forestal
#> 38                                                                             Institucional
#> 39                                                            Lote (Unidad sin construcción)
#> 40                                                                         Parque/Zona Verde
#> 41                                                                         Minero-Energético
#> 42                                                         Protección/Conservación ambiental
#> 43                                                                           En Construcción
#> 44                                                                           Sin información
#> 45                                                                                      <NA>
#> 46                                                                                      Casa
#> 47                                                                               Apartamento
#> 48                                                                               Tipo cuarto
#> 49                                                             Vivienda tradicional Indígena
#> 50                                Vivienda tradicional étnica (Afrocolombiana, Isleña, Rrom)
#> 51                Otro (contenedor, carpa, embarcación, vagón, cueva, refugio natural, etc.)
#> 52                                                            Ocupada con personas presentes
#> 53                                                   Ocupada con todas las personas ausentes
#> 54                                        Vivienda  temporal (para vacaciones, trabajo etc.)
#> 55                                                                                Desocupada
#> 56                                                                                      <NA>
#> 57                                                                                        Si
#> 58                                                                                        No
#> 59                                                                                 Estrato 1
#> 60                                                                                 Estrato 2
#> 61                                                                                 Estrato 3
#> 62                                                                                 Estrato 4
#> 63                                                                                 Estrato 5
#> 64                                                                                 Estrato 6
#> 65                                                                No sabe o no tiene estrato
#> 66                                                                                        Si
#> 67                                                                                        No
#> 68                                                                                        Si
#> 69                                                                                        No
#> 70                                                                                        Si
#> 71                                                                                        No
#> 72                                                                           Sin información
#> 73                                                                                        Si
#> 74                                                                                        No
#> 75                                                                                        Si
#> 76                                                                                        No
#> 77                                                                           Sin información
#> 78                                                                                      <NA>
#> 79                                                                                      <NA>
#> 80                                                                                      <NA>
#> 81                                                                                    Hombre
#> 82                                                                                     Mujer
#> 83                                                                                0 - 9 años
#> 84                                                                              10 - 19 años
#> 85                                                                              20 - 29 años
#> 86                                                                              30 - 39 años
#> 87                                                                              40 - 49 años
#> 88                                                                              50 - 59 años
#> 89                                                                              60 - 69 años
#> 90                                                                              70 - 79 años
#> 91                                                                            80  y más años
#> 92                                                   Preescolar - PrejardinBásica primaria 1
#> 93                                          Básica secundaria 6Media técnica 10Normalista 10
#> 94                             Técnica profesional 1 añoTecnológica 1 añoUniversitario 1 año
#> 95                                        Especialización 1 añoMaestria 1 añoDoctorado 1 año
#> 96                                                                                   Ninguno
#> 97                                                                           Sin información
#> 98                                                                                      <NA>