Download climate data from geometry

Download climate data from stations contained in a geometry. This data is retrieved from local meteorological stations provided by IDEAM.

Usage

download_climate_geom(
  geometry,
  start_date,
  end_date,
  frequency,
  tags,
  aggregate = TRUE
)

Arguments

geometry

sf geometry containing the geometry for a given ROI. This geometry can be either a POLYGON or MULTIPOLYGON

start_date

character with the first date to consult in the format "YYYY-MM-DD"

end_date

character with the last date to consult in the format "YYYY-MM-DD" (Last available date is "2023-05-31")

frequency

character with the aggregation frequency. ("day", "week", "month" or "year")

tags

character containing climate tags to consult

aggregate

logical for data aggregation, if FALSE, returns the data from each individual station in the area. (Default = TRUE)

Value

data.frame with observations from the stations in the area

Examples

lat <- c(5.166278, 5.166278, 4.982247, 4.982247, 5.166278)
lon <- c(-75.678072, -75.327859, -75.327859, -75.678072, -75.678072)
polygon <- sf::st_polygon(x = list(cbind(lon, lat)))
geometry <- sf::st_sfc(polygon)
roi <- sf::st_as_sf(geometry)
tssm <- download_climate_geom(
  roi, "2021-11-14", "2021-11-20",
  "day", "TSSM_CON"
)
#> Warning: Station 26155220 is not available
#> Warning: Station 26155210 is not available
#> Warning: Station 26155230 is not available
#> Warning: Station 26157090 is not available
#> Warning: Station 26157100 is not available
#> Warning: Station 26157060 is not available
#> Warning: Station 26155160 is not available
#> Warning: Station 26157080 is not available
#> Warning: Station 26157070 is not available
#> Warning: Station 26150160 is not available
#> Warning: Station 26157020 is not available
#> Warning: Station 26160100 is not available
#> Warning: Station 2615500122 is not available
#> Warning: Station 26150030 is not available
#> Warning: Station 26150040 is not available
#> Warning: Station 26150050 is not available
#> Warning: Station 26150100 is not available
#> Warning: Station 26150110 is not available
#> Warning: Station 26155190 is not available
#> Warning: Station 26155090 is not available
#> Warning: Station 26155180 is not available
#> Warning: Station 26157150 is not available
#> Warning: Station 23020130 is not available
#> Warning: Station 26150340 is not available
#> Warning: Station 26150360 is not available
#> Warning: Station 26150170 is not available
#> Warning: Station 26150410 is not available
#> Warning: Station 26150420 is not available
#> Warning: Station 26150310 is not available
#> Warning: Station 26155060 is not available
#> Warning: Station 26150140 is not available
#> Warning: Station 26150200 is not available
#> Warning: Station 26150400 is not available
#> Warning: Station 26150130 is not available
#> Warning: Station 26150370 is not available
#> Warning: Station 26155070 is not available
#> Warning: Station 26150390 is not available
#> Warning: Station 26157010 is not available
#> Warning: Station 26157050 is not available
#> Warning: Station 26130160 is not available
#> Warning: Station 26130470 is not available
#> Warning: Station 26135270 is not available
#> Warning: Station 26150220 is not available
#> Warning: Station 26150280 is not available
#> Warning: Station 26150010 is not available
#> Warning: Station 26150290 is not available
#> Warning: Station 26155504 is not available
#> Warning: Station 26150380 is not available
#> Warning: Station 26155020 is not available
#> Warning: Station 26155502 is not available
#> Warning: Station 24025070 is not available
#> Warning: Station 26137030 is not available
#> Warning: Station 26150080 is not available
#> Warning: Station 26150260 is not available
#> Warning: Station 26157030 is not available
#> Warning: Station 26150210 is not available
#> Warning: Station 26155200 is not available
#> Warning: Station 26150090 is not available
#> Warning: Station 26150180 is not available
#> Warning: Station 26150350 is not available
#> Warning: Station 26150320 is not available
#> Warning: Station 26155030 is not available
#> Warning: Station 26155100 is not available
#> Warning: Station 23025502 is not available
#> Warning: Station 26150070 is not available
#> Warning: Station 26150250 is not available
#> Warning: Station 26150300 is not available
#> Warning: Station 26155050 is not available
#> Warning: Station 26157130 is not available
#> Warning: Station 26150190 is not available
#> Warning: Station 26155503 is not available
#> Warning: Station 2615500011 is not available
#> Warning: Station 26155010 is not available
#> Warning: Station 26150020 is not available
#> Warning: Station 2615500005 is not available
#> Warning: Station 2615500008 is not available
#> Warning: Station 2615700040 is not available
#> Warning: Station 2615500012 is not available
#> Warning: Station 2615700045 is not available
#> Warning: Station 2615700046 is not available
#> Warning: Station 2615500015 is not available
#> Warning: Station 2615700048 is not available
#> Warning: Station 2615500021 is not available
#> Warning: Station 2615700053 is not available
#> Warning: Station 2615700056 is not available
#> Warning: Station 2615700061 is not available
#> Warning: Station 2615500026 is not available
#> Warning: Station 2615700066 is not available
#> Warning: Station 2615500028 is not available
#> Warning: Station 2615700067 is not available
#> Warning: Station 2615500031 is not available
#> Warning: Station 2615500033 is not available
#> Warning: Station 2615700074 is not available
#> Warning: Station 2615700076 is not available
#> Warning: Station 2613700077 is not available
#> Warning: Station 2615700079 is not available
#> Warning: Station 2615700080 is not available
#> Warning: Station 2615700082 is not available
#> Warning: Station 2615500035 is not available
#> Warning: Station 2615500013 is not available
#> Warning: Station 2615700041 is not available
#> Warning: Station 2615700042 is not available
#> Warning: Station 2615700043 is not available
#> Warning: Station 2615700044 is not available
#> Warning: Station 2615500014 is not available
#> Warning: Station 2615500016 is not available
#> Warning: Station 2615700049 is not available
#> Warning: Station 2615500018 is not available
#> Warning: Station 2615700050 is not available
#> Warning: Station 2615500019 is not available
#> Warning: Station 2615500020 is not available
#> Warning: Station 2615500022 is not available
#> Warning: Station 2615500023 is not available
#> Warning: Station 2615700054 is not available
#> Warning: Station 2615700055 is not available
#> Warning: Station 2615500024 is not available
#> Warning: Station 2615500025 is not available
#> Warning: Station 2615700058 is not available
#> Warning: Station 2615700059 is not available
#> Warning: Station 2615500027 is not available
#> Warning: Station 2615700064 is not available
#> Warning: Station 2615700065 is not available
#> Warning: Station 2615500029 is not available
#> Warning: Station 2615700068 is not available
#> Warning: Station 2615700073 is not available
#> Warning: Station 2615500034 is not available
#> Warning: Station 2615700001 is not available
print(tssm)
#>         date TSSM_CON
#> 1 2021-11-14    16.60
#> 2 2021-11-15    16.13
#> 3 2021-11-16    16.23
#> 4 2021-11-17    17.00
#> 5 2021-11-18    15.88
#> 6 2021-11-19    13.84
#> 7 2021-11-20    15.59