Download climate data from stations

Download climate data from IDEAM stations by individual codes.This data is retrieved from local meteorological stations provided by IDEAM.

Usage

download_climate_stations(stations, start_date, end_date, tag)

Arguments

stations

data.frame containing the stations' codes and location. data.frame must be retrieved from the function stations_in_roi()

start_date

character with the first date to consult in the format "YYYY-MM-DD". (First available date is "1920-01-01").

end_date

character with the last date to consult in the format "YYYY-MM-DD". (Last available date is "2023-05-31").

tag

character containing climate tag to consult.

Value

data.frame object with observations from the stations in the area.

Examples

# \donttest{
lat <- c(4.172817, 4.172817, 4.136050, 4.136050, 4.172817)
lon <- c(-74.749121, -74.686169, -74.686169, -74.749121, -74.749121)
polygon <- sf::st_polygon(x = list(cbind(lon, lat)))
geometry <- sf::st_sfc(polygon)
roi <- sf::st_as_sf(geometry)
stations <- stations_in_roi(roi)
ptpm <- download_climate_stations(
  stations, "2022-11-14", "2022-11-20", "PTPM_CON"
)
#> ColOpenData provides open data derived from Departamento Administrativo
#> Nacional de Estadística (DANE), and Instituto de Hidrología,
#> Meteorología y Estudios Ambientales (IDEAM) but with modifications for
#> specific functional needs. These changes may alter the structure,
#> format, or content, meaning the data does not reflect the official
#> dataset. The package is developed independently, with no endorsement or
#> involvement from these institutions or any Colombian government body.
#> The authors of ColOpenData are not liable for how users utilize the
#> data, and users are responsible for any outcomes from their use or
#> analysis of the data.
#> Stored by Universidad de Los Andes under the Epiverse TRACE initiative.
head(ptpm)
#>    station    longitude   latitude       date     hour      tag value
#> 1 21190290 -74.71311111 4.16102778 2022-11-14 07:00:00 PTPM_CON     2
#> 2 21190290 -74.71311111 4.16102778 2022-11-15 07:00:00 PTPM_CON     0
#> 3 21190290 -74.71311111 4.16102778 2022-11-16 07:00:00 PTPM_CON     2
#> 4 21190290 -74.71311111 4.16102778 2022-11-17 07:00:00 PTPM_CON    43
#> 5 21190290 -74.71311111 4.16102778 2022-11-18 07:00:00 PTPM_CON     0
#> 6 21190290 -74.71311111 4.16102778 2022-11-19 07:00:00 PTPM_CON     0
# }