Download data dictionaries

Retrieve geospatial data dictionaries to understand internal tags and named columns. Dictionaries are available in English and Spanish.

Usage

geospatial_dictionary(spatial_level, language = "ES")

Arguments

spatial_level

character with the spatial level to be consulted:

"DPTO" or "department": Department.

"MPIO" or "municipality": Municipality.

"MPIOCL" or "municipality_class": Municipality including class.

"SETU" or "urban_sector": Urban Sector.

"SETR" or "rural_sector": Rural Sector.

"SECU" or "urban_section": Urban Section.

"SECR" or "rural_section": Rural Section.

"ZU" or "urban_zone": Urban Zone.

"MZN" or "block": Block.

language

character with the language of the dictionary variables ("EN" or "ES". Default is "ES".

Value

data.frame object with geospatial data dictionary.

Examples

dict <- geospatial_dictionary("setu", "EN")
head(dict)
#> # A tibble: 6 × 4
#>   variable                 type  length description                             
#>   <chr>                    <chr>  <dbl> <chr>                                   
#> 1 codigo_departamento      Text       2 Department code                         
#> 2 codigo_municipio_sin_con Text       3 Municipality code                       
#> 3 codigo_municipio         Text       5 Concatenated municipality code          
#> 4 codigo_clase             Text       1 Class code 1 municipal seat, 2 populate…
#> 5 codigo_setr_sin_con      Text       3 Rural sector code                       
#> 6 codigo_setr              Text       9 Concatenated rural sector code (departm…