Filter list of available datasets based on keywords given by the user

List available datasets containing user-specified keywords in their descriptions.

Usage

look_up(keywords, module = "all", logic = "or", language = "EN")

Arguments

keywords

character or vector of characters to be look up in the description.

module

character with module to be consulted ("demographic", "geospatial", "climate"). Default is "all".

logic

A character string specifying the matching logic. Can be either "or" or "and". Default is "or":

logic = "or": Matches rows containing at least one of the specified keywords in their descriptions.

logic = "and": Matches rows containing all of the specified keywords in their descriptions.

language

character with the language of the keywords ("EN" or "ES". Default is "EN".

Value

data.frame object with the available datasets containing information related to the consulted keywords.

Examples

found <- look_up(c("sex", "age"), "demographic", "and", "EN")
head(found)
#> # A tibble: 6 × 7
#>   name                  group       source year  level      category description
#>   <chr>                 <chr>       <chr>  <chr> <chr>      <chr>    <chr>      
#> 1 DANE_CNPVPD_2018_1PD  demographic DANE   2018  department persons… Total cens…
#> 2 DANE_CNPVPD_2018_1PM  demographic DANE   2018  municipal… persons… Total cens…
#> 3 DANE_CNPVPD_2018_2PT  demographic DANE   2018  national   persons… Total cens…
#> 4 DANE_CNPVPD_2018_3PD  demographic DANE   2018  department persons… Total cens…
#> 5 DANE_CNPVPD_2018_3PM  demographic DANE   2018  municipal… persons… Total cens…
#> 6 DANE_CNPVPD_2018_5BPD demographic DANE   2018  department persons… Census pop…