Stations in region of interest

Download and filter climate stations contained inside a region of interest (ROI).

Usage

stations_in_roi(geometry)

Arguments

geometry

sf object containing the geometry for a given ROI. The geometry can be either a POLYGON or MULTIPOLYGON.

Value

data.frame object with the stations contained inside the consulted geometry.

Examples

# \donttest{
lat <- c(5.166278, 5.166278, 4.982247, 4.982247, 5.166278)
lon <- c(-75.678072, -75.327859, -75.327859, -75.678072, -75.678072)
polygon <- sf::st_polygon(x = list(cbind(lon, lat)))
geometry <- sf::st_sfc(polygon)
roi <- sf::st_as_sf(geometry)
stations <- stations_in_roi(roi)
head(stations)
#> Simple feature collection with 6 features and 20 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -75.51667 ymin: 5.029778 xmax: -75.35611 ymax: 5.1
#> CRS:           NA
#>        codigo                           nombre           categoria
#> 138  26155220      VILLAMARIA - AUT [26155220] Climática Principal
#> 139  26155210         TORRE 4 - AUT [26155210] Climática Principal
#> 140  26155230            EMAS - AUT [26155230] Climática Principal
#> 266  26155110   AEROPUERTO LA NUBIA [26155110] Sinóptica Principal
#> 749  26155170 TESORITO FINCA  - AUT [26155170] Climática Ordinaria
#> 1119 26157090              PLAYA LA [26157090]        Limnimétrica
#>                     tecnologia     estado departamento  municipio    latitud
#> 138  Automática con Telemetría     Activa       Caldas Villamaria 5.04866667
#> 139  Automática con Telemetría Suspendida       Caldas  Manizales 5.06077778
#> 140  Automática con Telemetría     Activa       Caldas  Manizales    5.08525
#> 266               Convencional     Activa       Caldas  Manizales 5.02977778
#> 749  Automática con Telemetría     Activa       Caldas  Manizales 5.03222222
#> 1119              Convencional Suspendida       Caldas  Manizales        5.1
#>          longitud altitud fecha_instalacion
#> 138  -75.51388889    1906        2004-06-14
#> 139  -75.35611111    3787        2004-07-14
#> 140  -75.50713889    2211        2004-06-10
#> 266  -75.46991667    2104        1968-06-15
#> 749  -75.43833333    2325  1992-10-14 19:00
#> 1119 -75.51666667     250        1984-07-15
#>                              area_operativa         corriente area_hidrografica
#> 138  Area Operativa 09 - Cauca-Valle-Caldas          San Juan   Magdalena Cauca
#> 139  Area Operativa 09 - Cauca-Valle-Caldas          San Juan   Magdalena Cauca
#> 140  Area Operativa 09 - Cauca-Valle-Caldas          San Juan   Magdalena Cauca
#> 266  Area Operativa 09 - Cauca-Valle-Caldas           Orinoco   Magdalena Cauca
#> 749  Area Operativa 09 - Cauca-Valle-Caldas                 0   Magdalena Cauca
#> 1119 Area Operativa 09 - Cauca-Valle-Caldas Quebrada Olivares   Magdalena Cauca
#>      zona_hidrografica subzona_hidrografica
#> 138              Cauca        Río Chinchiná
#> 139    Medio Magdalena          Río Guarinó
#> 140              Cauca        Río Chinchiná
#> 266              Cauca        Río Chinchiná
#> 749              Cauca        Río Chinchiná
#> 1119             Cauca        Río Chinchiná
#>                                                          entidad
#> 138  INSTITUTO DE HIDROLOGIA METEOROLOGIA Y ESTUDIOS AMBIENTALES
#> 139  INSTITUTO DE HIDROLOGIA METEOROLOGIA Y ESTUDIOS AMBIENTALES
#> 140  INSTITUTO DE HIDROLOGIA METEOROLOGIA Y ESTUDIOS AMBIENTALES
#> 266  INSTITUTO DE HIDROLOGIA METEOROLOGIA Y ESTUDIOS AMBIENTALES
#> 749  INSTITUTO DE HIDROLOGIA METEOROLOGIA Y ESTUDIOS AMBIENTALES
#> 1119 INSTITUTO DE HIDROLOGIA METEOROLOGIA Y ESTUDIOS AMBIENTALES
#>      fecha_suspension codigo_municipio codigo_departamento
#> 138              <NA>            17873                  17
#> 139        2010-09-29            17446                  17
#> 140              <NA>            17001                  17
#> 266              <NA>            17001                  17
#> 749              <NA>            17001                  17
#> 1119       1996-04-15            17001                  17
#>                        geometry
#> 138  POINT (-75.51389 5.048667)
#> 139  POINT (-75.35611 5.060778)
#> 140   POINT (-75.50714 5.08525)
#> 266  POINT (-75.46992 5.029778)
#> 749  POINT (-75.43833 5.032222)
#> 1119      POINT (-75.51667 5.1)
# }