Download geospatial dataset

This function downloads geospatial datasets from the Geostatistical National Framework at different levels of spatial aggregation. These datasets include a summarized version of the National Population and Dwelling Census (CNPV) with demographic and socioeconomic information for each spatial unit.

Usage

download_geospatial(spatial_level, include_geom = TRUE, include_cnpv = TRUE)

Arguments

spatial_level

character with the spatial level to be consulted

"DPTO" or "department": Department

"MPIO" or "municipality": Municipality

"MPIOCL" or "municipality_class": Municipality including class

"SETU" or "urban_sector": Urban Sector

"SETR" or "rural_sector": Rural Sector

"SECU" or "urban_section": Urban Section

"SECR" or "rural_section": Rural Section

"ZU" or "urban_zone": Urban Zone

"MZN" or "block": Block

include_geom

logical for including (or not) the geometry. Default is TRUE

include_cnpv

logical for including (or not) CNPV demographic and socioeconomic information. Default is TRUE

Value

data.frame object with downloaded data

Examples

departments <- download_geospatial("department", TRUE, FALSE)
#> Original data is retrieved from the National Administrative Department
#> of Statistics (Departamento Administrativo Nacional de Estadística -
#> DANE).
#> Reformatted by package authors.
#> Stored by Universidad de Los Andes under the Epiverse TRACE iniative.
head(departments)
#> Simple feature collection with 6 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -77.92834 ymin: -0.70584 xmax: -66.84722 ymax: 6.324317
#> Geodetic CRS:  WGS 84
#>   codigo_departamento    departamento version         area   latitud  longitud
#> 1                  18         Caquetá    2018  90103008160 0.7985562 -73.95947
#> 2                  19           Cauca    2018  31242914793 2.3968339 -76.82423
#> 3                  86        Putumayo    2018  25976283135 0.4522600 -75.85591
#> 4                  76 Valle del Cauca    2018  20665544525 3.8588583 -76.51869
#> 5                  94         Guainía    2018  71289354481 2.7278429 -68.81661
#> 6                  99         Vichada    2018 100063370591 4.7135571 -69.41400
#>                             geom
#> 1 MULTIPOLYGON (((-74.89423 2...
#> 2 MULTIPOLYGON (((-76.45922 3...
#> 3 MULTIPOLYGON (((-76.6705 1....
#> 4 MULTIPOLYGON (((-77.2381 4....
#> 5 MULTIPOLYGON (((-67.67638 3...
#> 6 MULTIPOLYGON (((-67.80972 6...