Filter list of available datasets based on keywords given by the user

List available datasets containing user-specified keywords in their descriptions

Usage

look_up(module = "all", keywords, logic = "or")

Arguments

module

character with module to be consulted ("demographic", "geospatial", "climate"). Default is "all"

keywords

character or vector of characters to be look up in the description

logic

A character string specifying the matching logic. Can be either "or" or "and". Default is "or"

logic = "or": Matches rows containing at least one of the specified keywords in their descriptions.

logic = "and": Matches rows containing all of the specified keywords in their descriptions.

Value

data.frame object with the available datasets

Examples

found <- look_up("demographic", c("sex", "age"), "and")
head(found)
#>                     name       group source year        level
#> 17  DANE_CNPVPD_2018_1PD demographic   DANE 2018   department
#> 18  DANE_CNPVPD_2018_1PM demographic   DANE 2018 municipality
#> 19  DANE_CNPVPD_2018_2PT demographic   DANE 2018     national
#> 20  DANE_CNPVPD_2018_3PD demographic   DANE 2018   department
#> 21  DANE_CNPVPD_2018_3PM demographic   DANE 2018 municipality
#> 26 DANE_CNPVPD_2018_5BPD demographic   DANE 2018   department
#>               category
#> 17 persons_demographic
#> 18 persons_demographic
#> 19 persons_demographic
#> 20 persons_demographic
#> 21 persons_demographic
#> 26 persons_demographic
#>                                                                                                                                description
#> 17                                    Total census population, by department, area, age group, masculinity and femininity indexes, and sex
#> 18                                  Total census population, by municipality, area, age group, masculinity and femininity indexes, and sex
#> 19                                                        Total census population in 2018, 2005, 1993, 1985 and 1973, by age group and sex
#> 20                                                                    Total census population, by department, age group, age, area and sex
#> 21                                                                  Total census population, by municipality, age group, age, area and sex
#> 26 Census population in particular households, by relationship or kinship to the head of household, by department, sex, area and age group