Match and merge geospatial and demographic datasets

This function downloads a demographic and a geospatial dataset, does a pivot to the demographic dataset based on a column of interest and merges with a geospatial dataset according to the level of the demographic dataset. Since the smallest level of spatial aggregation present in the demographic datasets is municipality, this functions can only match with the municipality and department code of geospatial datasets.

Usage

merge_geo_demographic(demographic_dataset, column)

Arguments

demographic_dataset

character with the demographic dataset name. Please use list_datasets("demographic") to check available datasets

column

character with the column we are interested in. It depends on the demographic dataset, so it is important to have checked the dataset beforehand

Value

data.frame object with the merged data

Examples

census_file <- "DANE_CNPVV_2018_9VD"
merged <- merge_geo_demographic(census_file, "tipo_de_servicio_sanitario")
#> Original data is retrieved from the National Administrative Department
#> of Statistics (Departamento Administrativo Nacional de Estadística -
#> DANE).
#> Reformatted by package authors.
#> Stored by Universidad de Los Andes under the Epiverse TRACE iniative.